# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
    "/blog/": {
      "get": {
        "tags": ["blog"],
        "summary": "Get All Blog Posts",
        "description": "Retrieve all blog posts.",
        "operationId": "get_all_blog_posts_blog__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/BlogPost" },
                  "type": "array",
                  "title": "Response Get All Blog Posts Blog  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["blog"],
        "summary": "Create Blog Post",
        "description": "Create a new blog post.",
        "operationId": "create_blog_post_blog__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BlogPostData" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BlogPost" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/blog/{blog_id}": {
      "get": {
        "tags": ["blog"],
        "summary": "Get Blog Post By Id",
        "description": "Retrieve a blog post by ID.",
        "operationId": "get_blog_post_by_id_blog__blog_id__get",
        "parameters": [
          {
            "name": "blog_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Blog Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BlogPost" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["blog"],
        "summary": "Update Blog Post",
        "description": "Update an existing blog post.",
        "operationId": "update_blog_post_blog__blog_id__put",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "blog_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Blog Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BlogPostData" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BlogPost" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["blog"],
        "summary": "Delete Blog Post",
        "description": "Delete a blog post by ID.",
        "operationId": "delete_blog_post_blog__blog_id__delete",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "blog_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Blog Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BlogPost": {
        "properties": {
          "title": { "type": "string", "title": "Title" },
          "slug": { "type": "string", "title": "Slug" },
          "content": { "type": "string", "title": "Content" },
          "date": { "type": "string", "title": "Date" },
          "id": { "type": "string", "format": "uuid", "title": "Id" }
        },
        "type": "object",
        "required": ["title", "slug", "content", "date", "id"],
        "title": "BlogPost"
      },
      "BlogPostData": {
        "properties": {
          "title": { "type": "string", "title": "Title" },
          "slug": { "type": "string", "title": "Slug" },
          "content": { "type": "string", "title": "Content" },
          "date": { "type": "string", "title": "Date" }
        },
        "type": "object",
        "required": ["title", "slug", "content", "date"],
        "title": "BlogPostData"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "login/access-token" }
        }
      }
    }
  }
}